+++ app/src/Navigation/index.js
@@ -1,5 +1,5 @@
 import React, { useContext } from 'react';
-import BottomTab from './MainNavigation/BottomNavigation';
+import BottomTab from './MainNavigation/BottomNavigator';
 import AuthRoutes from './AuthNavigation';
 import { NavigationContainer } from '@react-navigation/native';
 import { UserContext } from '../store/contexts/userContext';
